Ensure frozen dims are known when evaluating b_abs
This fixes issue with interaction with cache.

If variable identifier F is a b_abs with frozen dimensions, the
evaluation of F in different contexts must lead to several cache
entries for F.  E.g. if a b_abs named F has frozen dimension t, the
following two entries could end up in the cache:
* An evaluated b_abs when t=46
* An evaluated b_abs when t=1

Of course, the dimensions needed for evaluating the expressions
representing the frozen dimensions must be known as well in the first
place.