Close functions, and evaluate args in the correct environment
Evaluate args in the correct environment:
* Evaluate arguments in application environment/context - and not in
  closed/modified environment/context (as it was wrongly done
  before). This fixes a lot of tests;
* Remove '@' sugar for applications, re-introducing parallel
  evaluation of actual arguments and abstraction.

Close functions:
* Close the shallowest abstractions in the new expressions defined in
  wherevar clauses;
* Close unnamed abstractions that are not shallowest abstractions for
  any wherevar clauses;
* The elements that are stored in the cache are frozen closures;
* Adapt logic re freezing and abs/apply;
* The logic for closing expressions is in the new tclosure module;
* Add relevant negative tests.

The tests re closure of the dimension id expression in a tuple AST
node require the sequential version of the evaluator in order to catch
the error in the right process.  In order to do so, meck is used for
mocking tpar only for the tests needing the sequential evaluator.