In ttransform1, do not freeze local dims defined in outer wheredim clauses...
... unless explicitly frozen.

This is different from the Aug 2012 contextual semantics paper
(section 6.4, page 22), as the local dims defined in wheredim clauses
do not have the level of indirection (phi) defined in the Aug 2012
paper, they rather follow the model defined in the Feb 2013 cache
semantics paper.

Also update test suite. In particular, delete test groups aiming to
test if value functions are broken as they are not needed anymore.