Get rid of the cache's clock T
The ice:eval/1 function preserves the {term(), T} return value in
order not to update most tests at the moment.

Unit tests fail.  Further thought needs to be spent on this change as
actually the {term(), T} type in the core code (ice_core, ...) was
used for stopping execution as soon as possible with a badmatch in
case of cache hangs (and other errors maybe).